add snapshots
