+++ package.json
@@ -3,8 +3,7 @@
   "scripts": {
     "sync": "node sync-idl.js",
     "get-mints": "node ./scripts/fetch-mints-by-update-authority.js",
-    "parse-mints": "node ./scripts/parse-mints.js",
-    "fetch-listings": "ts-node ./scripts/fetch-listings.ts"
+    "snapshot": "ts-node ./scripts/snapshot.ts"
   },
   "dependencies": {
     "@project-serum/anchor": "^0.24.2",
