+++ scripts/snapshots/2022-06-24T15:56:24.017Z
@@ -1,0 +1,242 @@
+[
+  {
+    "publicKey": "99hNDaokzeWyEcwGiJrY8pL3Db9jL5DiNxTBjWjudPnr",
+    "account": {
+      "state": 4,
+      "amount": "b2d05e00",
+      "borrower": "9PweN2wakP3hwenNpqBTdoczokj331ZomgUVUwgLxvF8",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 5000,
+      "duration": "278d00",
+      "startDate": "00",
+      "escrow": "2CjgGtTeXTDPM9AenZJ1XQuZ6XErrvj31A1Q3t1aUe54",
+      "mint": "5ZK7cXRkE8FX31uXnkoVpgWn8VJaCF4YH5jHznwxTz14",
+      "bump": 254,
+      "escrowBump": 255
+    }
+  },
+  {
+    "publicKey": "6m8SouvG1wCV4CfXKY6tHrFtj2A9aG1djPYy1drwLLhv",
+    "account": {
+      "state": 1,
+      "amount": "1850aba0",
+      "borrower": "DCdL4k76N8NVaddpo2FhvkeFS1ZZiebQqHFtZfdDyN8s",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 16000,
+      "duration": "278d00",
+      "startDate": "00",
+      "escrow": "EPX8tEaMDU4Cki5Wg9Rwu4iMSM93anVKLnsYaDLXKL4C",
+      "mint": "4U3xNBu14DfrFuhZsAC9YetFaP4Kyo2vZoS1KZPBr5Yn",
+      "bump": 255,
+      "escrowBump": 255
+    }
+  },
+  {
+    "publicKey": "D6ksLRBbM9HNubg9ikj164jLEExzd6c75ptt5yuUKfiD",
+    "account": {
+      "state": 5,
+      "amount": "17d78400",
+      "borrower": "ErDUVBfMTeCAW2wPpXLqHDMTGzGKJVWNdD9HABjWFAdR",
+      "lender": "GadrUMEEnK8sw2HvpdbgHNt2ZZxh3GVmWUxYSndTWgkP",
+      "basisPoints": 5000,
+      "duration": "278d00",
+      "startDate": "622cea02",
+      "escrow": "2Bn3SKNNZLo1TUkXGfSAdhWw3Bm7sa5n47xwonJeJArP",
+      "mint": "2ffLAVpLubnH8AVaDTiAHkEwCvucCvmUkDRjw9hFyMQz",
+      "bump": 254,
+      "escrowBump": 255
+    }
+  },
+  {
+    "publicKey": "JBtipR3LqA3Wn3gp8NGFi8UudQUsXwLZ4e1t4nM6AZrg",
+    "account": {
+      "state": 1,
+      "amount": "1ed86c60",
+      "borrower": "3HxqsUguP6E7CNqjvpEAnJ8v86qbyJgWvN2idAKygLdD",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 24100,
+      "duration": "127500",
+      "startDate": "00",
+      "escrow": "2MFkmxoJs56WVCk5MdQ7LhB3QvbNZA4N52rm43MXRY9D",
+      "mint": "EBe6uhFfQ7hbYbtMihVtvq4QuKimeaeTBJwbovZHdH9m",
+      "bump": 255,
+      "escrowBump": 255
+    }
+  },
+  {
+    "publicKey": "BZvJxiPcQHXeQxvJ3MaGS7Y1BCgxuN1sdyDLu4jPsz6X",
+    "account": {
+      "state": 1,
+      "amount": "3b9aca00",
+      "borrower": "CpmYKMWogcXYv3wXZXthMkPzyYAy6n1fECbHvvyNn522",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 500,
+      "duration": "278d00",
+      "startDate": "00",
+      "escrow": "F6GxUnaDi4NzXGEyJLzyAzwzs1XmKmxV4C7ebE2y6wTi",
+      "mint": "BKsqyFeDbYxaTxYtMT5Nxw2gFi1rXngNBzTyD1CDhSWV",
+      "bump": 255,
+      "escrowBump": 255
+    }
+  },
+  {
+    "publicKey": "AGjnQAhBbkcMbY5iAUhoyHCBbznwskXEg6vqVtvX3MN9",
+    "account": {
+      "state": 4,
+      "amount": "77359400",
+      "borrower": "HvXB75H6GbfEMaCog31KHXH9o53AyhyQ3yH83Aqmktpr",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 500,
+      "duration": "278d00",
+      "startDate": "00",
+      "escrow": "AESAhN2zxFHdHhGdae2gNwoSiFGMwsUa3PCDDN3N7H7B",
+      "mint": "G8Vjqw1A4Mfv3dsrZUUWvhUTjHkJE7RuZEKZBtH9qz29",
+      "bump": 255,
+      "escrowBump": 251
+    }
+  },
+  {
+    "publicKey": "XgDC6vM4amAZi57bVWxUQ3yqj9v2PW6JvsXt9hWvBP7",
+    "account": {
+      "state": 4,
+      "amount": "17d78400",
+      "borrower": "FPtfUD2ZyfMroE44x9knzuGX4RrbPbA1HrSWHqDfpm69",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 3000,
+      "duration": "4f1a00",
+      "startDate": "00",
+      "escrow": "Gu8PLRZJT4S59annYspvGQGbG7AKmYzwzqDA59rcvjoV",
+      "mint": "8UbNHQS5cC9oRMA6d7BGRAxg7He1YPFHZgkwZ3j3B5Fq",
+      "bump": 255,
+      "escrowBump": 255
+    }
+  },
+  {
+    "publicKey": "CCMmmA9CzRnyeLeq3z6vhbr2RaTaSTyj7ujt4HswjK2d",
+    "account": {
+      "state": 4,
+      "amount": "3b9aca00",
+      "borrower": "9PweN2wakP3hwenNpqBTdoczokj331ZomgUVUwgLxvF8",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 5000,
+      "duration": "278d00",
+      "startDate": "00",
+      "escrow": "5TnQM6Hsrq84o98UJF5cXpybQ6fpKFGESwqHUCArxKCH",
+      "mint": "4q5P68ZRR4xjDJBLy9uGobyu6RjSq2FZwKpBMf9Pxuho",
+      "bump": 253,
+      "escrowBump": 254
+    }
+  },
+  {
+    "publicKey": "6Y2EEeq43f7HjdtA1FhsBHs4nspw1HjnLuWEFgxmoCtM",
+    "account": {
+      "state": 5,
+      "amount": "1dcd6500",
+      "borrower": "Dym7e6brxe47KEfoykg4SyXf6PhHNVFVYZh37VdXHtEC",
+      "lender": "3HxqsUguP6E7CNqjvpEAnJ8v86qbyJgWvN2idAKygLdD",
+      "basisPoints": 500,
+      "duration": "278d00",
+      "startDate": "621e6cf8",
+      "escrow": "CjaJ4Z7JDywZTrNUxf4ZW31ge314PFVo5dbtyv6fqKpa",
+      "mint": "FqBfyk4v6V4iXXdR1qfNhu4HbVKDFWNNeuVq6uL5vYp7",
+      "bump": 255,
+      "escrowBump": 255
+    }
+  },
+  {
+    "publicKey": "CV1AnF7KTCiud4bVRr4EmbQQnh9FwNiJtiDvv5EzT5Lm",
+    "account": {
+      "state": 4,
+      "amount": "1dcd6500",
+      "borrower": "AjoZvW6b2PJSNDDPhKVrtGHp8vubk7ubLbobcKrg4puT",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 500,
+      "duration": "278d00",
+      "startDate": "00",
+      "escrow": "AaqnR1hFQ8mXTLwD6mxPe8obfHeANwccC86Z4F9aS3LZ",
+      "mint": "66qG3oJeVpBAGe1GtDP22ENFNDzBpF78JgMtwodQcPHg",
+      "bump": 253,
+      "escrowBump": 254
+    }
+  },
+  {
+    "publicKey": "7o8ZAM3JKJivAW63SC7s7pDWNN6gmjkHhvYhmjTTaYAD",
+    "account": {
+      "state": 5,
+      "amount": "17d78400",
+      "borrower": "ErDUVBfMTeCAW2wPpXLqHDMTGzGKJVWNdD9HABjWFAdR",
+      "lender": "3HxqsUguP6E7CNqjvpEAnJ8v86qbyJgWvN2idAKygLdD",
+      "basisPoints": 5000,
+      "duration": "278d00",
+      "startDate": "6261ca89",
+      "escrow": "FsK1xtr6TMadAhbQ5bB2KC9zEFCE11ydHdWiyJhkhW21",
+      "mint": "46Ku8vQQ1Xfxey9duhJfvpa22PnzMhJ2ygxmh3SmDs5M",
+      "bump": 255,
+      "escrowBump": 252
+    }
+  },
+  {
+    "publicKey": "F192JZ4rhiiFfDhfpvQ4dnWeeCz8CECMYNpNE5LMrB7J",
+    "account": {
+      "state": 3,
+      "amount": "0bebc200",
+      "borrower": "7cN6LwUBQZwrHchFVyVRqKAtCzoMnJZUj1qrU6CW9u5o",
+      "lender": "3HxqsUguP6E7CNqjvpEAnJ8v86qbyJgWvN2idAKygLdD",
+      "basisPoints": 500,
+      "duration": "76a700",
+      "startDate": "621bf034",
+      "escrow": "FcwxqvzdxsTjuni47dVxcAaDPwtUMVaKCAb8jgPVm3aX",
+      "mint": "DQJ88T9uJ9pzymZnL5pnwASkLDttg5hR7b14Cyf1Z84b",
+      "bump": 254,
+      "escrowBump": 255
+    }
+  },
+  {
+    "publicKey": "DohUKUHGsLiMMCB2TLRpHkEWTFY8dLVw73BricX7KY5u",
+    "account": {
+      "state": 1,
+      "amount": "47868c00",
+      "borrower": "AH7F2EPHXWhfF5yc7xnv1zPbwz3YqD6CtAqbCyE9dy7r",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 5000,
+      "duration": "278d00",
+      "startDate": "00",
+      "escrow": "6qZgCkFvZWTiAFTxPoKW7rtW5SLHhjyHjhhuFXma9Hgf",
+      "mint": "Dh8osnrpjg3KXSDDVSvskb9vo5APtvZssRSCmQNeV4Nn",
+      "bump": 255,
+      "escrowBump": 255
+    }
+  },
+  {
+    "publicKey": "36E4YrfHhamivrNhyjBprNaqdkmje2J5S2tXrrrSg48h",
+    "account": {
+      "state": 1,
+      "amount": "04a817c800",
+      "borrower": "9QqRkp9uVod5tsbP3aYd86rakuLj2Wnxokrf3Y9WZwXd",
+      "lender": "11111111111111111111111111111111",
+      "basisPoints": 500,
+      "duration": "278d00",
+      "startDate": "00",
+      "escrow": "9TfPK5fwAbZE2qBKaSci1HH7s24nvwdRU1tF5Mzfik14",
+      "mint": "8UW2DfYx8zsYDnb3neESpnQoDh7Non34XK5S4NgATXTs",
+      "bump": 255,
+      "escrowBump": 253
+    }
+  },
+  {
+    "publicKey": "A4AgmCgxZ2ae9KZh6pmNx9JtyCtqQvuQAfFyhitp5fcg",
+    "account": {
+      "state": 5,
+      "amount": "17d78400",
+      "borrower": "HJBVadEgXB5TB3V65vn6gYvUvfeXtTzEW67kiR245Grf",
+      "lender": "3HxqsUguP6E7CNqjvpEAnJ8v86qbyJgWvN2idAKygLdD",
+      "basisPoints": 5000,
+      "duration": "278d00",
+      "startDate": "6261ca14",
+      "escrow": "C6DdLZM9mkaC4EuGDCMF9tVxzQQxA2WdDAYSxdpFSDvJ",
+      "mint": "EygEwX1WFStfXz2z1P1Fgx4JHJrYe6WW4hiLeC5qF6Yz",
+      "bump": 255,
+      "escrowBump": 253
+    }
+  }
+]
